+++ tony/Week2/Mission/Mission1/src/App.tsx
@@ -1,14 +1,13 @@
-import './App.css'
+import './index.css'
 import Todo from './components/Todo';
+import { TodoProvider } from './context/TodoContext';
+
 function App() {
-  
-  return (
-    <>
-    
-    <Todo />  
-    
-    </>
-  );
+    return (
+        <TodoProvider>
+            <Todo />
+        </TodoProvider>
+    );
 }
 
 export default App;
